Bus pass printing
A project developed for the Tamilnadu state transport Corporation
(TNSTC) using which free bus passes can be printed with reduced cost
than the already existing software’s. It is very simple software that
takes in the student details as input and renders the details into a
card like format.
Technology used: Microsoft Excel (Macros in VBA)